.examples/server-django-ninja/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohappyeyeballs==2.4.4",
    "aiohttp==3.11.11",
    "aiosignal==1.3.2",
    "annotated-types==0.7.0",
    "asgiref==3.8.1",
    "async-timeout==5.0.1",
    "attrs==25.1.0",
    "bcrypt==4.3.0",
    "cachetools==5.5.2",
    "cbor2==5.6.5",
    "cerberus==1.3.7",
    "certifi==2024.12.14",
    "charset-normalizer==3.4.1",
    "django==5.2.1",
    "django-cors-headers==4.7.0",
    "django-ninja==1.4.1",
    "dnspython==2.7.0",
    "email-validator==2.2.0",
    "frozenlist==1.5.0",
    "google-api-core==2.24.2",
    "google-api-python-client==2.169.0",
    "google-auth==2.40.1",
    "google-auth-httplib2==0.2.0",
    "google-auth-oauthlib==1.2.2",
    "googleapis-common-protos==1.70.0",
    "gunicorn>=23.0.0",
    "httplib2==0.22.0",
    "idna==3.10",
    "marshmallow==3.26.0",
    "multidict==6.1.0",
    "neo4j==5.28.1",
    "neomodel==5.4.5",
    "numpy==2.2.5",
    "oauthlib==3.2.2",
    "packaging==24.2",
    "pillow==11.2.1",
    "propcache==0.2.1",
    "proto-plus==1.26.1",
    "protobuf==6.30.2",
    "psycopg2-binary==2.9.10",
    "pyasn1==0.6.1",
    "pyasn1-modules==0.4.2",
    "pydantic==2.11.4",
    "pydantic-core==2.33.2",
    "pyjwt==2.10.1",
    "pyparsing==3.2.3",
    "python-dotenv==1.1.0",
    "pytz==2024.2",
    "requests==2.32.3",
    "requests-oauthlib==2.0.0",
    "rsa==4.9.1",
    "sqlparse==0.5.3",
    "stripe==12.1.0",
    "surrealdb==1.0.3",
    "typing-extensions==4.12.2",
    "typing-inspection==0.4.0",
    "uritemplate==4.1.1",
    "urllib3==2.3.0",
    "uvicorn>=0.35.0",
    "websockets==14.2",
    "whitenoise>=6.9.0",
    "yarl==1.18.3",
]
